move spacy readers to one file instead of directory
