--- ml_datasets/spacy_readers/__init__.py
@@ -1,3 +1,0 @@
-from .imdb_reader import imdb_reader
-from .cmu_reader import cmu_reader
-from .dbpedia_reader import dbpedia_reader
